TG_Bot_for_OP/requirements.txt
# Список залежностей
pyTelegramBotAPI==4.14.0
python-dotenv==1.0.0


